feat(CustomSeparatorHandler): 객체 생성에 의한 Separator 역할 분리
- CustomSeparatorHandler으로 메소드 이동
  커스텀 분리자 추출 메소드
  커스텀 분리자 정의 제거 메소드
